improved dist build
Added in an index.dist.html to dev folder and copy that to dist as
index.html to get around having to manually update base href when
copying dist contents to server.
Simplified gulpfile.js tasks so that it simply copies files from the dev
build.
Fixed .map files not being included in dist build which caused warnings
in chrome dev tools.